SEDAR/backend/commons/requirements.txt
APScheduler==3.7.0
delta-spark==1.0.0
Flask==2.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Werkzeug==2.0.2
flask-mongoengine==1.0.0
Flask-RESTful==0.3.9
kafka-python==2.0.2
passlib==1.7.4
psycopg2-binary
pyspark==3.1.2
pywebhdfs==0.4.1
tzlocal==2.1
python-dotenv==0.15.0
pydeequ3==0.1.8
neomodel==4.0.8
tika==1.24
elasticsearch==7.16.2
pycron==3.0.0
docker
